fix(app): PR 리뷰 반영 — 추천칩 카테고리행 통합, PlaceTags 사이즈/enum 갱신, queryKey+에러처리
- SearchCategory에 추천 칩 인라인 통합 (화장실 다음, 음식점 앞)
- PlaceSearchRecommendationChips 독립 컴포넌트 삭제 (ChipsOverlay 오버레이 제거)
- queryKey에 centerLocation 좌표 포함 (stale 방지)
- 추천 API 에러 시 ToastUtils.show 피드백 추가
- PlaceTags: SavedList → PlaceList enum 갱신, height 20 → padding-vertical 4px + font 10 → 12px (접근레벨 배지 높이 일치)
